requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
weaviate-client==3.25.3
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv

